[penge] Translate the times used in the event tile to local time
A JanaTime points to a particular point of time in universal time. Therefore
in order to find the time of an event in the local timezone the offset of the
time can be changed from the time that the event originates in to the local
timezone. This causes the internal time details to be updated appropriately
and birds will sing and the world will be a great place for one and all.