Fix handling of Panel-related windows in empty workspace check (MB#7162)
When a window closes that is related to a panel closes (e.g. Carrick Add New
Connection dialogue), we need to skip the empty workspace check, otherwise
the Panel will be forcefully closed if there are no applications running.
Also, we should not try to show the Myzone, if the some other Panel, or the
Toolbar are already visible.