Fix position of notifications, MB#1437
Fixed up the preferred size and allocate vfunctions to work in a more sensible
fashion, so that we can rely on the size of the allocation box for positioning
the notifications.

When watching for allocation changes, we need to query the allocation property,
as clutter_actor_get_size() is return spurious values.