[MnbSystray] MxBoxLayout subclass that sits on the top of ShellTrayManager
Automatically adds/removes children in response to the manager signals. This
actor *must* be hidden when not meant to be visible, and cannot be stacked
below anything in the Shell overlay when not hidden, as the X windows
associated with the tray icons, though transparent, always sit above the
ClutterStage window, and so will intercept mouse events.